Create FactorsOfNumber.java
This is a simple program to check whether a number is prime or not in O(√n) time. I use a counter to count the number of factors. If at any time the count becomes greater than two, it means the number is not prime. Since we know a prime number is a number that is divisible only by 1 and itself, the counter should remain 2.