Sieve of Eratosthenes.java
This is an algorithm used for finding prime numbers in a given range.

Implementation:
It is quite simple. We take a boolean array of size n+1 (to access the n-th element). Initially, all values are set to false, meaning they are assumed prime.

We then start a loop from 2 to √n (square root of n). We begin from 2 because it is the first prime number. For each i, if sieve[i] is still false (meaning it is prime), we mark all its multiples as true (non-prime). This works because a prime number has only two factors: 1 and itself. By marking multiples as true, we ensure that only prime numbers remain false.

The reason the outer loop runs only up to √n is that any non-prime number greater than √n must have a factor less than or equal to √n, which would have already been marked earlier.

TC= O(n log logn);